KP-6187 moved cta tag creation from user to activity

+++ activity/src/main/java/com/kairos/persistence/repository/tag/TagMongoRepository.java
@@ -48,6 +48,6 @@ public interface TagMongoRepository extends MongoRepository<Tag,BigInteger>,Cust
 
     Tag findTagByOrganizationIdAndNameAndMasterDataTypeAndDeletedAndCountryTagFalse(Long orgId, String name, String masterDataTypeEnum,boolean deleted);
 
-    List<TagDTO> findAllTagsByIdIn(List<Long> tagIds);
+    List<TagDTO> findAllTagsByIdIn(List<BigInteger> tagIds);
 }
